plugins/ril: add function ril_sim_get_state
ril_sim_get_state is called by qcommsimmodem's radio settings
driver. It is used to resolve problem with 3G slot selection.
Normally, to set a slot as 3G when there are 2 SIMs in the device,
radio settings must set other slot to 2G first. But if the other
slot is empty, this action will fail. So, radio settings must know
if the slot is empty, and skip setting this slot to 2G.