qcommsimmodem: don't set 2G radio preference if the slot is empty.
Normally, to set a slot as 3G when there are 2 SIMs in the device,
radio settings must set other slot to 2G first. But if the other
slot is empty, this action will fail. This commit make sure that
radio settings will skip the empty slot. Along the way, I've
refactored code around so that the code is cleaner.